requirements.txt
numpy>=2.4.2
scipy>=1.17.0
pillow>=12.1.0
